Moved connection.py to package folder
